data_pipeline/mint_addr/requirements.txt
# Requirements for Solana Mint Address Scraper
# Install with: pip install -r requirements.txt

# Async HTTP requests (required for API scraping)
aiohttp>=3.8.0

# Data processing and CSV handling
pandas>=1.5.0
numpy>=1.24.0

# Date/time utilities
python-dateutil>=2.8.0

# Configuration management
pyyaml>=6.0

# Optional: Solana SDK (only needed for blockchain scraper)
# solana-py>=0.30.0
# solders>=0.18.0

# Optional: Progress bars
tqdm>=4.64.0

# Optional: Better logging
colorlog>=6.7.0
